grr
svn:sync-xref-src-repo-rev: r85866

+++ src/VBox/Runtime/common/dbg/dbgmoddwarf.cpp
@@ -1,4 +1,4 @@
-/* $Id: dbgmoddwarf.cpp 46165 2013-05-19 19:07:50Z knut.osmundsen@oracle.com $ */
+/* $Id: dbgmoddwarf.cpp 46166 2013-05-19 19:48:20Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - Debug Info Reader For DWARF.
  */
@@ -4415,7 +4415,7 @@ static DECLCALLBACK(int) rtDbgModDwarfEnumCallback(RTLDRMOD hLdrMod, PCRTLDRDBGI
 static int rtDbgModDwarfTryOpenDbgFile(PRTDBGMODINT pDbgMod, PRTDBGMODDWARF pThis, RTLDRARCH enmArch)
 {
     if (   !pDbgMod->pszDbgFile
-        || RTPathIsSame(pDbgMod->pszDbgFile, pDbgMod->pszImgFile) == true /* returns VERR too */)
+        || RTPathIsSame(pDbgMod->pszDbgFile, pDbgMod->pszImgFile) == (int)true /* returns VERR too */)
         return VERR_DBG_NO_MATCHING_INTERPRETER;
 
     /*
